Use int32 in Go interfaces that map to protobufs using int32
We use plain `int`, which can be 32 or 64 bit based on the system, in some of the Go interfaces that map to the protobuf interface, but the corresponding protobuf interfaces always int32.

Update the Go interfaces to use int32 to prevent potential integer overflow.

+++ sdk/go/common/resource/plugin/langruntime_plugin.go
@@ -227,7 +227,7 @@ func (h *langhost) Run(info RunInfo) (string, bool, error) {
 		ConfigPropertyMap: configPropertyMap,
 		DryRun:            info.DryRun,
 		QueryMode:         info.QueryMode,
-		Parallel:          int32(info.Parallel),
+		Parallel:          info.Parallel,
 		Organization:      info.Organization,
 		Info:              minfo,
 	})
